Refactor index module into index::{in_memory,builder} and add crate-local constructor
